move graphqlRequests into central file
